requirements.txt
pandas
soccerdata
gspread
google-auth
flask
gunicorn
